pyproject.toml
[project]
classifiers = []
dependencies = ["psutil>=7.2.2", "pyobjc-framework-osakit>=12.1"]
description = "Lists the wallpaper file path currently showing on each desktop/space on macOS."
keywords = []
readme = "README.md"
license = { text = "MIT" }
name = "wpls"
version = "0.1.0"
requires-python = ">=3.10"
[[project.authors]]
name = "AJ Slater"
email = "aj@slater.net"

[project.urls]
News = "https://github.com/ajslater/wpls/NEWS.md"
Documentation = "https://github.com/ajslater/wpls/README.md"
Issues = "https://github.com/ajslater/wpls/issues"
Source = "https://github.com/ajslater/wpls"

[project.scripts]
wpls = "wpls:main"

[dependency-groups]
dev = [
  "infer-types~=1.0.0",
  "neovim~=0.3",
  "semver~=3.0.4",
  "toml-cli~=0.7",
  "tomlkit~=0.14",
]
build = []
lint = [
  "basedpyright~=1.38",
  "codespell~=2.4",
  "complexipy~=5.1",
  "icecream~=2.1",
  "mbake~=1.4.5",
  "pathspec~=1.0.4",
  "radon[toml]~=6.0",
  "ruff~=0.15",
  "types-python-dateutil~=2.9",
  "ty~=0.0.17",
  "vulture~=2.3",
]
test = [
  "coverage[toml]~=7.0",
  "pytest-asyncio~=1.0",
  "pytest-cov~=7.0",
  "pytest-gitignore~=1.3",
]
docs = ["mkdocs-minify-plugin~=0.8", "mkdocs~=1.6"]
ci = ["requests~=2.32"]

[build-system]
requires = ["uv_build~=0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
source-include = [
  ".*ignore",
  ".circlci/**",
  ".env.platforms",
  ".picopt_treestamps.yaml",
  ".readthedocs.yaml",
  ".remarkignore",
  ".shellcheckrc",
  "Makefile",
  "NEWS.md",
  "bin/**",
  "ci/**",
  "compose.yaml",
  "docs/DOCKER.md",
  "docs/NEWS",
  "docs/NEWS.md",
  "docs/README.md",
  "docs/WINDOWS.md",
  "docs/requirements.txt",
  "docs/style.css",
  "docs/style.material.css",
  "docs/style.mkdocs.css",
  "docs/style.readthedocs.css",
  "eslint.config.js",
  "frontend/**",
  "img/**",
  "mkdocks.yml",
  "mock_comics/**",
  "package-lock.json",
  "package.json",
  "tests/**",
  "uv.lock",
  "vulture_ignorelist.py",
]
source-exclude = ["**/*\\~", "**/.DS_Store", "frontend/node_modules"]
wheel-exclude = ["**/*\\~", "**/.DS_Store"]

[tool.basedpyright]
exclude = [
  "**/.*",
  "**/__pycache__",
  "**/node_modules",
  "dist",
  "frontend",
  "site",
  "site",
  "test-results",
  "typings",
  "vulture_ignorelist.py",
]
pythonVersion = "3.10"
failOnWarnings = false
reportAny = false
reportExplicitAny = false
reportImportCycles = true
reportIncompatibleUnannotatedOverride = true
reportMissingParameterType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportPrivateUsage = false
reportUnannotatedClassAttribute = false
reportUnknownArgumentType = false
reportUnknownLambdaType = "hint"
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnusedCallResult = false
reportUnusedParameter = false

[tool.codespell]
builtin = "clear,code,rare"
check-hidden = true
ignore-words-list = "coverd,falsy,ro,searchd,thead,versio,ws"
skip = "*~,.*,./dist,./node_modules,./package-lock.json,./package.json,./test-results,/uv.lock"

[tool.complexipy]
failed = true
paths = ["tests"]
exclude = [
  "*/.*",
  "frontend",
  "node_modules",
  "site",
  "test-results",
  "typings",
  "vulture_ignorelist.py",
]

[tool.coverage.html]
directory = "test-results/coverage"

[tool.coverage.run]
source = []
branch = true
concurrency = ["multiprocessing"]
omit = [
  "*/.*",
  "*__pycache__*",
  "dist/*",
  "frontend/*",
  "node_modules/*",
  "site/*",
  "test-results/*",
  "typings/*",
  "vulture_ignorelist.py",
]

[tool.pytest]
minversion = "9.0"
addopts = [
  "--cov",
  "--cov-append",
  "--cov-report=html",
  "--cov-report=term",
  "--junit-xml=test-results/pytest/results.xml",
  "--strict-config",
  "--strict-markers",
  "-ra",
]
junit_family = "xunit2"
strict = true
testpaths = ["tests"]
pythonpath = ["."]

[tool.radon]
exclude = "*~,.*,.*/*,__pycache__/*,config/*,dist/*,frontend/*,node_modules/*,site/*,test-results/*,typings/*"

[tool.ruff]
builtins = ["ic"]
extend-exclude = [
  "**/.*",
  "**/__pycache__",
  "dist",
  "frontend",
  "site",
  "test-results",
  "typings",
  "vulture_ignorelist.py",
]
show-fixes = true
target-version = "py310"

[tool.ruff.lint]
extend-ignore = [
  "BLE001",
  "COM812",
  "COM819",
  "D203",
  "D206",
  "D212",
  "E111",
  "E114",
  "E117",
  "E501",
  "ISC001",
  "PERF203",
  "S101",
  "W191",
]
extend-select = [
  "A",
  "ARG",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "D",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FIX",
  "FLY",
  "FURB",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "LOG",
  "N",
  "PERF",
  "PGH",
  "PIE",
  "PLE",
  "PLR",
  "PLW",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
external = ["V101"]
task-tags = ["FIXME", "HACK", "TODO", "XXX", "http"]

[tool.ruff.lint.flake8-self]
ignore-names = ["_meta"]

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D101", "D102", "T203"]
"wpls/__init__.py" = ["T201"]
"wpls/list.py" = ["T201"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.ty.src]
exclude = [
  "**/.*",
  "**/__pycache__",
  "**/node_modules",
  "dist",
  "frontend",
  "site",
  "test-results",
  "typings",
  "vulture_ignorelist.py",
]

[tool.typos.default.extend-words]
bookmarkd = "bookmarkd"
coverd = "coverd"
versio = "versio"

[tool.vulture]
exclude = [
  "*/.*",
  "*/__pycache__*",
  "*/node_modules*",
  "dist/",
  "frontend/",
  "site",
  "test_results/",
  "typings/",
]
min_confidence = 61
sort_by_size = true
